% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_gender21_ltla21_ew}
\alias{age_gender21_ltla21_ew}
\title{Age and gender grouped data for England and Wales from 2021 Census}
\format{
A data frame of class "tbl" with 11 rows and 11 variables:
\describe{
\item{ltla21_code}{Local Authority code}
\item{total_population}{Total population}
\item{total_female_population}{Total female population}
\item{total_male_population}{Total male population}
\item{younger_females}{Count of females aged under 20 years}
\item{working_age_females}{Count of females aged between 20 and 65}
\item{older_females}{Count of females aged over 65}
\item{younger_males}{Count of males aged under 20 years}
\item{working_age_males}{Count of males aged between 20 and 65}
\item{older_males}{Count of males aged over 65}
}
}
\source{
\url{https://www.nomisweb.co.uk/sources/census_2021_bulk}
}
\usage{
age_gender21_ltla21_ew
}
\description{
A dataset containing Census 2021 age and gender count for each LA in England and Wales
}
\keyword{datasets}
