% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_gender23_ltla_ni}
\alias{age_gender23_ltla_ni}
\title{Age and gender grouped data for Northern Ireland from 2023 Mid Year Estimates
(Local Government Districts)}
\format{
A data frame of class "tbl" with 11 rows and 11 variables:
\describe{
\item{ltla21_name}{Local Government District name (2014). Renamed LTLA for consistency.}
\item{ltla21_code}{Local Government District code (2014). Renamed LTLA for consistency.}
\item{total_population}{Total population}
\item{total_female_population}{Total female population}
\item{total_male_population}{Total male population}
\item{younger_females}{Count of females aged under 16 years}
\item{working_age_females}{Count of females aged between 16 and 64}
\item{older_females}{Count of females aged over 64}
\item{younger_males}{Count of males aged under 16 years}
\item{working_age_males}{Count of males aged between 16 and 64}
\item{older_males}{Count of males aged over 64}
}
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
age_gender23_ltla_ni
}
\description{
A dataset containing 2023 mid year estimates age and gender count for each Local
Government District in Northern Ireland.
}
\keyword{datasets}
