% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-error.R
\name{bootstrap_replicates_parallel}
\alias{bootstrap_replicates_parallel}
\title{Parallel Bootstrap Results Using Replicate Weights}
\usage{
bootstrap_replicates_parallel(
  data,
  f,
  wt_col = "weight",
  repwt_cols = paste0("repwt", 1:80),
  id_cols,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{A tibble or data frame containing the data to be analyzed.}

\item{f}{A function that produces new columns or summary statistics.
The function \code{f} must have an argument named \code{wt_col} to specify the weight column.}

\item{wt_col}{A string indicating the column name to be used as the main weight.}

\item{repwt_cols}{A vector of strings indicating the names of replicate weight
columns in \code{data}.}

\item{id_cols}{A character vector of columns that uniquely identify rows in the output.}

\item{verbose}{If TRUE, prints progress updates.}

\item{...}{Additional arguments passed to the function \code{f}.}
}
\value{
A list with:
\describe{
\item{main_estimate}{The result of applying \code{f()} with the main weight column.}
\item{replicate_estimates}{A list of results from applying \code{f()} with each
replicate weight column (computed in parallel).}
}
}
\description{
Same as \code{bootstrap_replicates()} but uses parallel mapping to speed up replicate estimation.
Uses \code{furrr::future_map2()} and respects the current \code{future::plan()}.
}
