% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household-adjacency-matrices.R
\name{count_components}
\alias{count_components}
\title{Count connected components in household adjacency matrix}
\usage{
count_components(mat)
}
\arguments{
\item{mat}{Adjacency matrix}
}
\value{
igraph components object
}
\description{
Count connected components in household adjacency matrix
}
