% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_of_birth21_ltla21}
\alias{country_of_birth21_ltla21}
\title{Country of birth data, as of Census 2021}
\format{
A data frame of class "tbl" with 27142 rows and 6 variables:
\describe{
\item{ltla21_name}{Local Authority name}
\item{ltla21_code}{Local Authority code}
\item{total_all_usual_residents}{Total number of residents in that LTLA}
\item{country_of_birth}{Country of birth}
\item{value}{Number from each country}
\item{prop}{Proportion}
...
}
}
\source{
\url{https://www.nomisweb.co.uk/sources/census_2021_bulk}
}
\usage{
country_of_birth21_ltla21
}
\description{
A data sets containing Census 2021 data on the country of birth by Lower
Tier Local Authority
}
\keyword{datasets}
