% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disability21_lgd21_ni}
\alias{disability21_lgd21_ni}
\title{Disability data for Northern Ireland from the 2021 Census (LGD)}
\format{
A data frame of class "tbl":
\describe{
\item{lgd21_code}{LGD code}
\item{disability}{Disability status}
\item{count}{The count of people with the disability status}
\item{prop}{The percentage of people with the disability status}
...
}
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
disability21_lgd21_ni
}
\description{
A data LSOA containing Census 2021 disability numbers and percentages for
each LTLA in England and Wales.
}
\keyword{datasets}
