% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disability21_lsoa21_ew}
\alias{disability21_lsoa21_ew}
\title{Disability data for England and Wales from the 2021 Census (Lower Super Output
Area)}
\format{
A data frame of class "tbl":
\describe{
\item{lsoa21_code}{LSOA code}
\item{disability}{Disability status}
\item{n}{The count of people with the disability status}
\item{total_residents}{Total residents in LSOA}
\item{prop}{The percentage of people with the disability status}
...
}
}
\source{
\url{https://www.nomisweb.co.uk/sources/census_2021_bulk}
}
\usage{
disability21_lsoa21_ew
}
\description{
A dataset containing Census 2021 disability numbers and percentages for
each LSOA in England and Wales.
}
\keyword{datasets}
