% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disability21_sdz21_ni}
\alias{disability21_sdz21_ni}
\title{Disability data for Northern Ireland from the 2021 Census (Super Data Zones)}
\format{
A data frame of class "tbl":
\describe{
\item{sdz21_code}{SDZ code}
\item{total_households}{Total number of households in each SDZ}
\item{disability}{Disability status}
\item{n}{The count of households with the disability status}
\item{prop}{The percentage of households with the disability status}
...
}
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
disability21_sdz21_ni
}
\description{
A data SDZ containing Census 2021 disability numbers and percentages for
each SDZ in Northern Ireland.
}
\keyword{datasets}
