% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disability22_dz11_scot}
\alias{disability22_dz11_scot}
\title{Disability data for Scotland from the 2022 Census (Data Zones)}
\format{
A data frame of class "tbl":
\describe{
\item{dz11_code}{DZ code}
\item{total_residents}{Total number of residents in each DZ}
\item{disability}{Disability status}
\item{n}{The count of residents with the disability status}
\item{prop}{The percentage of residents with the disability status}
...
}
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
disability22_dz11_scot
}
\description{
A data DZ containing Census 2022 disability numbers and percentages for
each DZ in Scotland.
}
\keyword{datasets}
