% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ethnicity21_detailed_msoa21}
\alias{ethnicity21_detailed_msoa21}
\title{Ethnic group (detailed) data for England and Wales from the 2021 Census
(MSOA)}
\format{
A data frame of class "tbl" with 2,084,768 rows and 5 variables:
\describe{
\item{msoa21_code}{MSOA code}
\item{total_residents}{Total number of residents in the MSOA}
\item{ethnic_group}{Ethnic group (detailed)}
\item{n}{Number of people in the ethnic group}
\item{prop}{Percentage of people in the ethnic group}
...
}
}
\source{
\url{https://ons.gov.uk/}
}
\usage{
ethnicity21_detailed_msoa21
}
\description{
A data sets containing Census 2021 detailed ethnic group numbers and
percentages for each Middle Layer Super Output Area in England and Wales.
}
\keyword{datasets}
