% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ethnicity21_lsoa21}
\alias{ethnicity21_lsoa21}
\title{Ethnic group data for England and Wales from the 2021 Census (LSOA)}
\format{
A data frame of class "tbl" with 5 variables:
\describe{
\item{mloa21_code}{LSOA code}
\item{total_residents}{Total number of residents in the LSOA}
\item{ethnic_group}{Ethnic group}
\item{n}{Number of people in the ethnic group}
\item{prop}{Percentage of people in the ethnic group}
...
}
}
\source{
\url{https://ons.gov.uk/}
}
\usage{
ethnicity21_lsoa21
}
\description{
A data sets containing Census 2021 ethnic group numbers and percentages for
each Lower Layer Super Output Area in England and Wales.
}
\keyword{datasets}
