% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ethnicity21_ltla21}
\alias{ethnicity21_ltla21}
\title{Ethnic group data for England and Wales from the 2021 Census}
\format{
A data frame of class "tbl" with 331 rows and 3 variables:
\describe{
\item{ltla21_name}{Local Authority name}
\item{ltla21_code}{Local Authority code}
\item{ethnic_group}{Ethnic group}
\item{value}{The count or percentage of the ethnic group}
...
}
}
\source{
\url{https://ons.gov.uk/}
}
\usage{
ethnicity21_ltla21
}
\description{
A data sets containing Census 2021 ehtnic group numers and percentages for
each Lower Tier Local Authority in England and Wales.
}
\keyword{datasets}
