% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ethnicity21_ltla21_ni}
\alias{ethnicity21_ltla21_ni}
\title{Ethnic group data for Northern Ireland from the 2021 Census (Local Authority)}
\format{
A data frame of class "tbl" with 286 rows and 4 variables:
\describe{
\item{ltla21_name}{Local Authority name}
\item{ltla21_code}{Local Authority code}
\item{ethnic_group}{Ethnic group}
\item{value}{The count or percentage of the ethnic group}
...
}
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
ethnicity21_ltla21_ni
}
\description{
A data sets containing Census 2021 ethnic group numbers and percentages for
each Local Government District in Northern Ireland.
}
\keyword{datasets}
