% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household_deprivation11_dz11}
\alias{household_deprivation11_dz11}
\title{Deprivation dimensions for Data Zones in Scotland}
\format{
A data frame of class "tbl" with 34,880 rows and 5 variables:
\describe{
\item{dz11_code}{Data Zone code}
\item{total_households}{Total number of households in each DZ}
\item{deprivation_dimensions}{Deprived in dimensions (0 to 4)}
\item{n}{Number of households per each deprivation dimension}
\item{prop}{The percentage of households with the deprivation dimension status}
}
}
\source{
\url{https://www.scotlandscensus.gov.uk/search-the-census}
}
\usage{
household_deprivation11_dz11
}
\description{
A data set containing 2011 estimates of deprivation dimensions for Data Zones in Scotland.
}
\keyword{datasets}
