% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household_deprivation21_oa21}
\alias{household_deprivation21_oa21}
\title{Deprivation dimensions for Output Areas in England and Wales}
\format{
A data frame of class "tbl" with 944,400 rows and 5 variables:
\describe{
\item{oa21_code}{Output Area code}
\item{total_households}{Total number of households in each OA}
\item{deprivation_dimensions}{Deprived in dimensions (0 to 4)}
\item{n}{Number of households per each deprivation dimension}
\item{prop}{The percentage of households with the deprivation dimension status}
}
}
\source{
\url{https://www.nomisweb.co.uk/sources/census_2021_bulk}
}
\usage{
household_deprivation21_oa21
}
\description{
A data set containing 2021 estimates of deprivation dimensions for Output Areas in England and Wales.
}
\keyword{datasets}
