% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household_deprivation21_sdz21}
\alias{household_deprivation21_sdz21}
\title{Deprivation dimensions for Super Data Zones in Northern Ireland}
\format{
A data frame of class "tbl" with 4,250 rows and 5 variables:
\describe{
\item{sdz21_code}{Super Data Zone code}
\item{total_households}{Total number of households in each SDZ}
\item{deprivation_dimensions}{Deprived in dimensions (0 to 4)}
\item{n}{Number of households per each deprivation dimension}
\item{prop}{The percentage of households with the deprivation dimension status}
}
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
household_deprivation21_sdz21
}
\description{
A data set containing 2021 estimates of deprivation dimensions for Super Data Zones in Northern Ireland.
}
\keyword{datasets}
