% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household_overcrowding21_sdz21}
\alias{household_overcrowding21_sdz21}
\title{Household overcrowding statistics for Super Data Zones in Northern Ireland}
\format{
A data frame of class "tbl" with 4,250 rows and 5 variables:
\describe{
\item{sdz21_code}{Super Data Zone code}
\item{total_households}{Total number of households in each SDZ}
\item{occupancy_rating}{Occupancy rating}
\item{n}{Number of households per each occupancy rating}
\item{prop}{The percentage of households with the specific occupancy rating}
}
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
household_overcrowding21_sdz21
}
\description{
A data set containing 2021 estimates of household overcrowding for Super Data Zones in Northern Ireland.
}
\keyword{datasets}
