% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household_overcrowding22_dz11}
\alias{household_overcrowding22_dz11}
\title{Household overcrowding statistics for Data Zones in Scotland}
\format{
A data frame of class "tbl" with 27,904 rows and 5 variables:
\describe{
\item{dz11_code}{Data Zone code}
\item{total_households}{Total number of households in each DZ}
\item{occupancy_rating}{Occupancy rating}
\item{n}{Number of households per each occupancy rating}
\item{prop}{The percentage of households with the specific occupancy rating}
}
}
\source{
\url{https://www.scotlandscensus.gov.uk/scotlands-census-2022-flexible-table-builder/}
}
\usage{
household_overcrowding22_dz11
}
\description{
A data set containing 2022 estimates of household overcrowding for Data Zones in Scotland.
}
\keyword{datasets}
