% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household_tenure21_lsoa21}
\alias{household_tenure21_lsoa21}
\title{Household tenure statistics for Lower Super Output Areas in England and Wales}
\format{
A data frame of class "tbl" with 285,376 rows and 5 variables:
\describe{
\item{lsoa21_code}{Lower Super Output Area code}
\item{total_households}{Total number of households in each LSOA}
\item{tenure}{Household tenure type}
\item{n}{Number of households per each tenure type}
\item{prop}{The percentage of households with the specific tenure type}
}
}
\source{
\url{https://www.nomisweb.co.uk/sources/census_2021_bulk}
}
\usage{
household_tenure21_lsoa21
}
\description{
A data set containing 2021 estimates of household tenure for Lower Super Output Areas in England and Wales.
}
\keyword{datasets}
