% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household_tenure21_oa21}
\alias{household_tenure21_oa21}
\title{Household tenure statistics for Output Areas in England and Wales}
\format{
A data frame of class "tbl" with 1,511,040 rows and 5 variables:
\describe{
\item{oa21_code}{Output Area code}
\item{total_households}{Total number of households in each OA}
\item{tenure}{Household tenure type}
\item{n}{Number of households per each tenure type}
\item{prop}{The percentage of households with the specific tenure type}
}
}
\source{
\url{https://www.nomisweb.co.uk/sources/census_2021_bulk}
}
\usage{
household_tenure21_oa21
}
\description{
A data set containing 2021 estimates of household tenure for Output Areas in England and Wales.
}
\keyword{datasets}
