% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household_tenure22_dz11}
\alias{household_tenure22_dz11}
\title{Household tenure statistics for Data Zones in Scotland}
\format{
A data frame of class "tbl" with 55,808 rows and 5 variables:
\describe{
\item{dz11_code}{Data Zone code}
\item{total_households}{Total number of households in each DZ}
\item{tenure}{Household tenure type}
\item{n}{Number of households per each tenure type}
\item{prop}{The percentage of households with the specific tenure type}
}
}
\source{
\url{https://www.scotlandscensus.gov.uk/scotlands-census-2022-flexible-table-builder/}
}
\usage{
household_tenure22_dz11
}
\description{
A data set containing 2022 estimates of household tenure for Data Zones in Scotland.
}
\keyword{datasets}
