% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household_type21_lsoa21}
\alias{household_type21_lsoa21}
\title{Household type statistics for Lower Super Output Areas in England and Wales}
\format{
A data frame of class "tbl" with 285,376 rows and 5 variables:
\describe{
\item{lsoa21_code}{Lower Super Output Area code}
\item{total_households}{Total number of households in each LSOA}
\item{type}{Household type}
\item{n}{Number of households per each household type}
\item{prop}{The percentage of households with the specific household type}
}
}
\source{
\url{https://www.nomisweb.co.uk/sources/census_2021_bulk}
}
\usage{
household_type21_lsoa21
}
\description{
A data set containing 2021 estimates of household type for Lower Super Output Areas in England and Wales.
}
\keyword{datasets}
