% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household_type21_oa21}
\alias{household_type21_oa21}
\title{Household type statistics for Output Areas in England and Wales}
\format{
A data frame of class "tbl" with 1,511,040 rows and 5 variables:
\describe{
\item{oa21_code}{Output Area code}
\item{total_households}{Total number of households in each OA}
\item{type}{Household type}
\item{n}{Number of households per each household type}
\item{prop}{The percentage of households with the specific household type}
}
}
\source{
\url{https://www.nomisweb.co.uk/sources/census_2021_bulk}
}
\usage{
household_type21_oa21
}
\description{
A data set containing 2021 estimates of household type for Output Areas in England and Wales.
}
\keyword{datasets}
