% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household_type21_sdz21}
\alias{household_type21_sdz21}
\title{Household type statistics for Super Data Zones in Northern Ireland}
\format{
A data frame of class "tbl" with 5,950 rows and 5 variables:
\describe{
\item{sdz21_code}{Super Data Zone code}
\item{total_households}{Total number of households in each SDZ}
\item{type}{Household type}
\item{n}{Number of households per each household type}
\item{prop}{The percentage of households with the specific household type}
}
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
household_type21_sdz21
}
\description{
A data set containing 2021 estimates of household type for Super Data Zones in Northern Ireland.
}
\keyword{datasets}
