% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household_type22_dz11}
\alias{household_type22_dz11}
\title{Household type statistics for Data Zones in Scotland}
\format{
A data frame of class "tbl" with 48,832 rows and 5 variables:
\describe{
\item{dz11_code}{Data Zone code}
\item{total_households}{Total number of households in each DZ}
\item{type}{Household type}
\item{n}{Number of households per each household type}
\item{prop}{The percentage of households with the specific household type}
}
}
\source{
\url{https://www.scotlandscensus.gov.uk/scotlands-census-2022-flexible-table-builder/}
}
\usage{
household_type22_dz11
}
\description{
A data set containing 2022 estimates of household type for Data Zones in Scotland.
}
\keyword{datasets}
