% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_lsoa_msoa}
\alias{lookup_lsoa_msoa}
\title{Lower Layer Super Output Area (2011) to Middle Layer Super Output Area (2011) Lookup}
\format{
A data frame of class "tbl" with 41,729 rows and 4 variables:
\describe{
\item{lsoa_name}{LSOA name}
\item{lsoa_code}{LSOA code}
\item{msoa_name}{MSOA name}
\item{msoa_code}{MSOA code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_lsoa_msoa
}
\description{
A data set containing a lookup between 2011 Lower (LSOA) and Middle Layer Super Output Areas (MSOA).
}
\keyword{datasets}
