% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{migration21_msoa21}
\alias{migration21_msoa21}
\title{A data sets containing Census 2021 migration statistics with percentages for
each Middle Super Output Area (MSOA).}
\format{
A data frame of class "tbl" with 7264 rows and 9 variables:
\describe{
\item{msoa21_code}{MSOA code}
\item{msoa21_name}{MSOA name}
\item{all_usual_residents}{Number of all usual residents}
\item{address_one_year_ago_is_the_same}{Number where address one year ago is the same as the address of enumeration}
\item{address_one_year_ago_is_student_or_boarding}{Number where address one year ago is student term-time or boarding school address in the UK}
\item{migrant_from_within_uk}{Migrant from within the UK: Number where address one year ago was in the UK}
\item{migrant_from_outside_uk}{Migrant from outside the UK: Number where address one year ago was outside the UK}
\item{prop_migrated_from_within_uk}{Proportion of migrants from within the UK}
\item{prop_migrated_from_outside_uk}{Proportion of migrants from outside the UK}
...
}
}
\source{
\url{https://www.nomisweb.co.uk/sources/census_2021_bulk}
}
\usage{
migration21_msoa21
}
\description{
A data sets containing Census 2021 migration statistics with percentages for
each Middle Super Output Area (MSOA).
}
\keyword{datasets}
