% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population17_lsoa11}
\alias{population17_lsoa11}
\title{Population estimates for LSOAs in England and Wales (mid-2017 estimates with 2011 codes)}
\format{
A data frame of class "tbl" with 34,753 rows and 94 variables:
\describe{
\item{lsoa11_name}{LSOA name}
\item{lsoa11_code}{LSOA code}
\item{total_population}{Total number of people living in each LSOA}
\item{\code{0}:\verb{90+}}{Number of people in each single age group}
...
}
}
\source{
\url{https://ons.gov.uk/}
}
\usage{
population17_lsoa11
}
\description{
A data set containing mid-2017 population estimates for Lower Layer Super Output Areas in England and Wales.
}
\keyword{datasets}
