% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population20_ltla20}
\alias{population20_ltla20}
\title{Population estimates for Local Authority Districts (mid-2020 values with 2020 codes)}
\format{
A data frame of class "tbl" with 336 rows and 94 variables:
\describe{
\item{ltla20_name}{LAD name}
\item{ltla20_code}{LAD code}
\item{total_population}{Total number of people living in each LAD}
\item{\code{0}:\verb{90+}}{Number of people in each single age group}
...
}
}
\source{
\url{https://ons.gov.uk/}
}
\usage{
population20_ltla20
}
\description{
A data set containing mid-2020 population estimates for Local Authority
Districts (April 2020 codes) in each devolved nation.
}
\keyword{datasets}
