% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population20_oa11}
\alias{population20_oa11}
\title{Population estimates for OAs in England and Wales (mid-2020 estimates with 2011 codes)}
\format{
A data frame of class "tbl" with 181,408 rows and 94 variables:
\describe{
\item{oa11_name}{OA name}
\item{oa11_code}{OA code}
\item{total_population}{Total number of people living in each OA}
\item{\code{0}:\verb{90+}}{Number of people in each single age group}
...
}
}
\source{
\url{https://ons.gov.uk/}
}
\usage{
population20_oa11
}
\description{
A data set containing mid-2020 population estimates for Output Areas in England and Wales.
}
\keyword{datasets}
