% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population20_soa11}
\alias{population20_soa11}
\title{Population estimates for SOAs in Northern Ireland (mid-2020 estimates with 2011 codes)}
\format{
A data frame of class "tbl" with 4,537 rows and 2 variables:
\describe{
\item{soa11_code}{SOA code}
\item{total_population}{Total number of people living in each SOA}
...
}
}
\source{
\url{https://nisra.gov.uk/}
}
\usage{
population20_soa11
}
\description{
A data set containing mid-2020 population estimates for Super Output Areas in Northern Ireland.
}
\keyword{datasets}
