% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population21_country21}
\alias{population21_country21}
\title{Population in UK countries, as of Census 2021}
\format{
A data frame of class "tbl" with 4 rows and 3 variables:
\describe{
\item{country_name}{Country name}
\item{country_code}{Country code}
\item{population}{Number of people}
...
}
}
\source{
\url{https://ons.gov.uk/}
}
\usage{
population21_country21
}
\description{
A data sets containing Census 2021 population counts for each UK nation.
Note that Scotland has not yet released its latest Census results, so this
is based on mid-year population estimates.
}
\keyword{datasets}
