% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population21_dea14}
\alias{population21_dea14}
\title{Usual Resident Population, District Electoral Area, Northern Ireland, based on Census 2021}
\format{
A data frame of class "tbl" with 80 rows and 3 variables:
\describe{
\item{dea14_name}{District Electoral Area name}
\item{dea14_code}{District Electoral Area code}
\item{residents}{All usual residents}
...
}
}
\source{
\url{https://www.nisra.gov.uk/publications/census-2021-person-and-household-estimates-data-zones-northern-ireland}
}
\usage{
population21_dea14
}
\description{
Usual Resident Population, District Electoral Area, Northern Ireland, based on Census 2021
}
\keyword{datasets}
