% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population21_dz11}
\alias{population21_dz11}
\title{Population estimates in Scotland's Data Zones (2011 codes), as of 2021}
\format{
A data frame of class "tbl" with 6,976 rows and 3 variables:
\describe{
\item{dz11_name}{Data Zone name}
\item{dz11_code}{Data Zone code}
\item{total_population}{Number of people}
...
}
}
\source{
\url{https://statistics.gov.scot/}
}
\usage{
population21_dz11
}
\description{
Population estimates in Scotland's Data Zones (2011 codes), as of 2021
}
\keyword{datasets}
