% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population21_dz21_ni}
\alias{population21_dz21_ni}
\title{Population in Data Zones, Northern Ireland, 2021}
\format{
A data frame of class "tbl"::
\describe{
\item{dz21_code}{Data Zone code}
\item{n_people}{Number of usual residents}
...
}
}
\source{
\url{https://www.nisra.gov.uk/publications/census-2021-person-and-household-estimates-data-zones-northern-ireland}
}
\usage{
population21_dz21_ni
}
\description{
A data set containing 2021 population counts
}
\keyword{datasets}
