% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population21_iz11}
\alias{population21_iz11}
\title{Population estimates in Scotland's Intermediate Zones (2011 codes), as of 2021}
\format{
A data frame of class "tbl" with 1,279 rows and 3 variables:
\describe{
\item{iz11_name}{Intermediate Zone name}
\item{iz11_code}{Intermediate Zone code}
\item{total_population}{Number of people}
...
}
}
\source{
\url{https://statistics.gov.scot/}
}
\usage{
population21_iz11
}
\description{
Population estimates in Scotland's Intermediate Zones (2011 codes), as of 2021
}
\keyword{datasets}
