% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population21_lgd14}
\alias{population21_lgd14}
\title{Usual Resident Population, Local Government District, Northern Ireland, based on Census 2021}
\format{
A data frame of class "tbl" with 11 rows and 3 variables:
\describe{
\item{lgd14_name}{Local Government District name}
\item{lgd14_code}{Local Government District code}
\item{residents}{All usual residents}
...
}
}
\source{
\url{https://www.nisra.gov.uk/publications/census-2021-person-and-household-estimates-data-zones-northern-ireland}
}
\usage{
population21_lgd14
}
\description{
Usual Resident Population, Local Government District, Northern Ireland, based on Census 2021
}
\keyword{datasets}
