% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population21_sdz21}
\alias{population21_sdz21}
\title{Usual Resident Population, Super Data Zone, Northern Ireland, based on Census 2021}
\format{
A data frame of class "tbl" with 850 rows and 3 variables:
\describe{
\item{sdz21_name}{Super Data Zone name}
\item{sdz21_code}{Super Data Zone code}
\item{residents}{All usual residents}
...
}
}
\source{
\url{https://www.nisra.gov.uk/publications/census-2021-person-and-household-estimates-data-zones-northern-ireland}
}
\usage{
population21_sdz21
}
\description{
Usual Resident Population, Super Data Zone, Northern Ireland, based on Census 2021
}
\keyword{datasets}
