% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population22_lsoa21}
\alias{population22_lsoa21}
\title{Population in England and Wales Lower Layer Super Output Areas, 2022 mid year estimates}
\format{
A data frame of class "tbl" with 25,672 rows and 2 variables:
\describe{
\item{lsoa21_code}{LSOA code}
\item{n_people}{Estimated number of people in LSOA}
...
}
}
\source{
\url{https://ons.gov.uk}
}
\usage{
population22_lsoa21
}
\description{
Population in England and Wales Lower Layer Super Output Areas, 2022 mid year estimates
}
\keyword{datasets}
