% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population22_ltla19_scotland}
\alias{population22_ltla19_scotland}
\title{Population in Scotland Councils, 2022 mid year estimates}
\format{
A data frame of class "tbl" with 32 rows and 95 variables:
\describe{
\item{ltla19_name}{Local Authority name}
\item{ltla19_code}{Local Authority code}
\item{sex}{Sex}
\item{all_ages}{Total number of people}
\item{\code{0}:\verb{90 and over}}{Number of people in each single age group}
...
}
}
\source{
\url{https://www.nrscotland.gov.uk/statistics-and-data/statistics/statistics-by-theme/population/population-estimates/mid-year-population-estimates/mid-2022#:~:text=Of\%20the\%2032\%20council\%20areas,and\%20Orkney\%20Islands\%20with\%2022\%2C020)}
}
\usage{
population22_ltla19_scotland
}
\description{
A data set containing 2022 mid year estimate population counts for each
Council in Scotland by sex, age.
}
\keyword{datasets}
