% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population22_oa21_ew}
\alias{population22_oa21_ew}
\title{Population in Output Areas, England and Wales, 2022 mid year estimates}
\format{
A data frame of class "tbl"::
\describe{
\item{oa21_code}{Output Area name}
\item{n_people}{Number of people}
...
}
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/censusoutputareapopulationestimatessupportinginformation}
}
\usage{
population22_oa21_ew
}
\description{
A data set containing 2022 mid year estimate population counts
}
\keyword{datasets}
