% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population22_oa22_s}
\alias{population22_oa22_s}
\title{Population in Output Areas, Scotland, 2022 mid year estimates}
\format{
A data frame of class "tbl"::
\describe{
\item{oa22_code}{Output Area code}
\item{n_people}{Number of people}
...
}
}
\source{
\url{https://www.nrscotland.gov.uk/publications/2022-census-geography-products/}
}
\usage{
population22_oa22_s
}
\description{
A data set containing 2022 mid year estimate population counts
}
\keyword{datasets}
