% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_counties_ua_19_codes_19}
\alias{population_counties_ua_19_codes_19}
\title{Population estimates for Counties and Unitary Authorities (mid-2019 with 2019 codes)}
\format{
A data frame of class "tbl" with 173 rows and 94 variables:
\describe{
\item{county_ua_name}{County or Unitary Authority name}
\item{county_ua_code}{County or Unitary Authority code}
\item{total_population}{Total number of people living in each area}
\item{\code{0}:\verb{90+}}{Number of people in each single age group}
...
}
}
\source{
\url{https://ons.gov.uk/}
}
\usage{
population_counties_ua_19_codes_19
}
\description{
A data set containing mid-2019 population estimates for County and Unitary
Authority Districts (April 2019 codes) in each devolved nation.
}
\keyword{datasets}
