% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_hb19}
\alias{population_hb19}
\title{Scottish Health Board (2019) Population Estimates}
\format{
A data frame of class "tbl" with 574 rows and 3 variables:
\describe{
\item{hb19_code}{NHS Health Board code}
\item{year}{Year of population estimate}
\item{total_population}{Total population for all ages}
...
}
}
\source{
\url{https://www.opendata.nhs.scot/}
}
\usage{
population_hb19
}
\description{
A data set containing population estimates for NHS Scotland Health Boards
(2019), from 1981 to 2021.
}
\keyword{datasets}
