% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_msoa_19}
\alias{population_msoa_19}
\title{Population estimates for MSOAs in England and Wales (mid-2019 estimates)}
\format{
A data frame of class "tbl" with 7,201 rows and 94 variables:
\describe{
\item{msoa_name}{MSOA name}
\item{msoa_code}{MSOA code}
\item{total_population}{Total number of people living in each MSOA}
\item{\code{0}:\verb{90+}}{Number of people in each single age group}
...
}
}
\source{
\url{https://ons.gov.uk/}
}
\usage{
population_msoa_19
}
\description{
A data set containing mid-2019 population estimates for Middle Layer Super Output Areas in England and Wales.
}
\keyword{datasets}
