% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{religion21_lsoa21}
\alias{religion21_lsoa21}
\title{Religion data for England and Wales from the 2021 Census (MSOA)}
\format{
A data frame of class "tbl" with 5 variables:
\describe{
\item{lsoa21_code}{LSOA code}
\item{total_residents}{Total number of residents in the LSOA}
\item{religion}{Religion}
\item{n}{Number of people in the religion}
\item{prop}{Percentage of people in the religion}
...
}
}
\source{
\url{https://ons.gov.uk/}
}
\usage{
religion21_lsoa21
}
\description{
A data sets containing Census 2021 religion numbers and percentages for
each Lower Layer Super Output Area in England and Wales.
}
\keyword{datasets}
