% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sexualorientation21_lgd21_ni}
\alias{sexualorientation21_lgd21_ni}
\title{Religion data for NI from the 2021 Census (LGD)}
\format{
A data frame of class "tbl" with 4 variables:
\describe{
\item{lgd21_code}{LGD code}
\item{sexual_orientation}{Sexual Orientation}
\item{count}{Number of people}
\item{prop}{Percentage of people}
...
}
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
sexualorientation21_lgd21_ni
}
\description{
A data sets containing Census 2021 sexual orientation numbers and percentages for
each Local Government District in Northern Ireland.
}
\keyword{datasets}
