% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sexualorientation21_msoa21_ew}
\alias{sexualorientation21_msoa21_ew}
\title{Religion data for England and Wales from the 2021 Census (MSOA)}
\format{
A data frame of class "tbl" with 4 variables:
\describe{
\item{msoa21_code}{MSOA code}
\item{sexual_orientation}{Sexual Orientation}
\item{count}{Number of people}
\item{prop}{Percentage of people}
...
}
}
\source{
\url{https://ons.gov.uk/}
}
\usage{
sexualorientation21_msoa21_ew
}
\description{
A data sets containing Census 2021 sexual orientation numbers and percentages for
each MSOA in England and Wales
}
\keyword{datasets}
