% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population21_dz11}
\alias{population21_dz11}
\title{Population estimates for Data Zones in Scotland (2021 estimates with 2011 codes)}
\format{
A data frame of class "tbl" with 20,928 rows and 94 variables:
\describe{
\item{dz11_code}{Data Zone code}
\item{total_population}{Total number of people living in each DZ}
\item{sex}{'Female', 'Male' or 'All'}
\item{\code{0}:\verb{90+}}{Number of people in each single age group}
...
}
}
\source{
\url{https://www.opendata.nhs.scot/}
}
\usage{
population21_dz11
}
\description{
A data set containing 2021 population estimates by age and sex for Data Zones in Scotland.
}
\keyword{datasets}
