% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population21_ltla21}
\alias{population21_ltla21}
\title{Population in UK Lower Tier Local Authorities, as of Census 2021}
\format{
A data frame of class "tbl" with 331 rows and 3 variables:
\describe{
\item{ltla21_name}{Local Authority name}
\item{ltla21_code}{Local Authority code}
\item{population}{Number of people}
...
}
}
\source{
\url{https://ons.gov.uk/}
}
\usage{
population21_ltla21
}
\description{
A data sets containing Census 2021 population counts for each Lower Tier
Local Authority in the UK. Currently only Local Authority data has been
released for England and Wales.
}
\keyword{datasets}
