% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{religion21_lsoa21}
\alias{religion21_lsoa21}
\title{Religion data for England and Wales from the 2021 Census (MSOA)}
\format{
A data frame of class "tbl" with 5 variables:
\describe{
\item{lsoa21_code}{LSOA code}
\item{total_residents}{Total number of residents in the LSOA}
\item{religion}{Religion}
\item{n}{Number of people in the religion}
\item{prop}{Percentage of people in the religion}
...
}

A data frame of class "tbl" with 5 variables:
\describe{
\item{lgd21_code}{LGD code}
\item{religion}{Religion}
\item{count}{Number of people}
\item{prop}{Percentage of people}
...
}
}
\source{
\url{https://ons.gov.uk/}

\url{https://www.nisra.gov.uk/}
}
\usage{
religion21_lsoa21

religion21_lsoa21
}
\description{
A data sets containing Census 2021 religion numbers and percentages for
each Lower Layer Super Output Area in England and Wales.

A data sets containing Census 2021 religion numbers and percentages for
each Local Government District in Northern Ireland.
}
\keyword{datasets}
